Added the send func and fixed a test bug.
The bug was an improper use of the context managers, as of Attest 0.5.
Currently copies are made of the context, so the context ends up not
being shared between tests. This caused the connection test to pass, but
other tests to fail because the connection was only valid within that
single connection test.